deleted added

+++ src/app/components/logs/logs.component.ts
@@ -29,7 +29,11 @@ console.log(log);
 this.logService.setFormLog(log);
 }
 onDelete(log: Log) {
-  this.logService.deleteLog(log);
+  if(confirm('Are you sure')){
+    this.logService.deleteLog(log);
+  }
+  
 }
+
 
 }
